# hotel-booking-api/app/Http/Controllers/Api/AdminDichVuController.php
<?php
namespace App\Http\Controllers\Api;
use App\Http\Controllers\Controller;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\DB;
use Illuminate\Support\Facades\Validator;
use Illuminate\Validation\Rule;
use OpenApi\Attributes as OA;
class AdminDichVuController extends Controller {
#[OA\Get(path:'/api/admin/dich-vu',summary:'Lấy danh sách dịch vụ quản trị',tags:['Quản trị - Dịch vụ'],security:[['sanctum'=>[]]],responses:[new OA\Response(response:200,description:'Thành công')])]
public function index(Request $request): JsonResponse { $v=Validator::make($request->query(),['tim_kiem'=>['nullable','string','max:255'],'page'=>['nullable','integer','min:1'],'per_page'=>['nullable','integer','min:1','max:50']]); if($v->fails()) return $this->loi($v->errors()); $d=$v->validated(); $p=DB::table('dich_vu')->when(!empty($d['tim_kiem']),fn($q)=>$q->where('ten_dich_vu','like','%'.$d['tim_kiem'].'%'))->orderByDesc('created_at')->paginate($d['per_page']??10); return response()->json(['success'=>true,'message'=>'Lấy danh sách dịch vụ thành công.','data'=>collect($p->items())->map(fn($x)=>$this->format($x)),'pagination'=>$this->pagination($p)],options:JSON_UNESCAPED_UNICODE); }
#[OA\Post(path:'/api/admin/dich-vu',summary:'Thêm dịch vụ',tags:['Quản trị - Dịch vụ'],security:[['sanctum'=>[]]],responses:[new OA\Response(response:201,description:'Thành công')])]
public function store(Request $r): JsonResponse { $v=Validator::make($r->all(),$this->rules(),$this->messages()); if($v->fails())return $this->loi($v->errors()); $id=DB::table('dich_vu')->insertGetId([...$v->validated(),'trang_thai'=>'hoat_dong','created_at'=>now(),'updated_at'=>now()]); return response()->json(['success'=>true,'message'=>'Thêm dịch vụ thành công.','data'=>$this->find($id)],201,options:JSON_UNESCAPED_UNICODE); }
#[OA\Get(path:'/api/admin/dich-vu/{id}',summary:'Lấy chi tiết dịch vụ',tags:['Quản trị - Dịch vụ'],security:[['sanctum'=>[]]],parameters:[new OA\Parameter(name:'id',in:'path',required:true,schema:new OA\Schema(type:'integer'))],responses:[new OA\Response(response:200,description:'Thành công'),new OA\Response(response:404,description:'Không tìm thấy dịch vụ')])]
public function show(int $id): JsonResponse { $x=$this->find($id); return $x?response()->json(['success'=>true,'message'=>'Lấy dịch vụ thành công.','data'=>$x],options:JSON_UNESCAPED_UNICODE):$this->notFound(); }
#[OA\Put(path:'/api/admin/dich-vu/{id}',summary:'Cập nhật dịch vụ',tags:['Quản trị - Dịch vụ'],security:[['sanctum'=>[]]],parameters:[new OA\Parameter(name:'id',in:'path',required:true,schema:new OA\Schema(type:'integer'))],responses:[new OA\Response(response:200,description:'Thành công'),new OA\Response(response:404,description:'Không tìm thấy dịch vụ'),new OA\Response(response:422,description:'Dữ liệu không hợp lệ')])]
public function update(Request $r,int $id): JsonResponse { if(!$this->find($id))return $this->notFound();$v=Validator::make($r->all(),$this->rules($id,true),$this->messages());if($v->fails())return $this->loi($v->errors());DB::table('dich_vu')->where('id',$id)->update([...$v->validated(),'updated_at'=>now()]);return response()->json(['success'=>true,'message'=>'Cập nhật dịch vụ thành công.','data'=>$this->find($id)],options:JSON_UNESCAPED_UNICODE); }
#[OA\Put(path:'/api/admin/dich-vu/{id}/ngung-hoat-dong',summary:'Ngừng hoạt động dịch vụ',tags:['Quản trị - Dịch vụ'],security:[['sanctum'=>[]]],parameters:[new OA\Parameter(name:'id',in:'path',required:true,schema:new OA\Schema(type:'integer'))],responses:[new OA\Response(response:200,description:'Thành công'),new OA\Response(response:404,description:'Không tìm thấy dịch vụ'),new OA\Response(response:422,description:'Không thể cập nhật trạng thái')])]
public function ngungHoatDong(int $id): JsonResponse{return $this->status($id,'ngung_hoat_dong','Đã ngừng hoạt động dịch vụ.');}
#[OA\Put(path:'/api/admin/dich-vu/{id}/kich-hoat',summary:'Kích hoạt lại dịch vụ',tags:['Quản trị - Dịch vụ'],security:[['sanctum'=>[]]],parameters:[new OA\Parameter(name:'id',in:'path',required:true,schema:new OA\Schema(type:'integer'))],responses:[new OA\Response(response:200,description:'Thành công'),new OA\Response(response:404,description:'Không tìm thấy dịch vụ'),new OA\Response(response:422,description:'Không thể cập nhật trạng thái')])]
public function kichHoat(int $id): JsonResponse{return $this->status($id,'hoat_dong','Đã kích hoạt lại dịch vụ.');}
private function status(int $id,string $s,string $m):JsonResponse{$x=$this->find($id);if(!$x)return $this->notFound();if($x->trang_thai===$s)return response()->json(['success'=>false,'message'=>'Trạng thái dịch vụ không thay đổi.'],422,options:JSON_UNESCAPED_UNICODE);DB::table('dich_vu')->where('id',$id)->update(['trang_thai'=>$s,'updated_at'=>now()]);return response()->json(['success'=>true,'message'=>$m,'data'=>$this->find($id)],options:JSON_UNESCAPED_UNICODE);}
private function rules(?int $id=null,bool $u=false):array{$r=$u?'sometimes':'required';return ['ten_dich_vu'=>[$r,'string','max:255',Rule::unique('dich_vu','ten_dich_vu')->ignore($id)],'mo_ta'=>[$u?'sometimes':'nullable','nullable','string'],'gia'=>[$r,'numeric','min:0']];} private function messages():array{return ['ten_dich_vu.required'=>'Tên dịch vụ là bắt buộc.','ten_dich_vu.unique'=>'Tên dịch vụ đã tồn tại.','gia.required'=>'Giá dịch vụ là bắt buộc.','gia.numeric'=>'Giá dịch vụ phải là số.','gia.min'=>'Giá dịch vụ không được nhỏ hơn 0.'];} private function find(int $id):?object{$x=DB::table('dich_vu')->where('id',$id)->first();return $x?$this->format($x):null;}private function format(object $x):object{$x->gia=(float)$x->gia;return $x;}private function pagination($p):array{return ['current_page'=>$p->currentPage(),'last_page'=>$p->lastPage(),'per_page'=>$p->perPage(),'total'=>$p->total()];}private function loi($e):JsonResponse{return response()->json(['success'=>false,'message'=>'Dữ liệu không hợp lệ.','errors'=>$e],422,options:JSON_UNESCAPED_UNICODE);}private function notFound():JsonResponse{return response()->json(['success'=>false,'message'=>'Không tìm thấy dịch vụ.','data'=>null],404,options:JSON_UNESCAPED_UNICODE);}
}