# hotel-booking-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking API",
        "description": "REST API cho hệ thống đặt phòng khách sạn.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/admin/dich-vu": {
            "get": {
                "tags": [
                    "Quản trị - Dịch vụ"
                ],
                "summary": "Lấy danh sách dịch vụ quản trị",
                "operationId": "4de847f2109c1cad5157e49159c109cd",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Quản trị - Dịch vụ"
                ],
                "summary": "Thêm dịch vụ",
                "operationId": "f9d5789ac7f98030975491814efbac35",
                "responses": {
                    "201": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dich-vu/{id}": {
            "get": {
                "tags": [
                    "Quản trị - Dịch vụ"
                ],
                "summary": "Lấy chi tiết dịch vụ",
                "operationId": "aa55119adaa26251ff45a4e40ccfea8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy dịch vụ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Quản trị - Dịch vụ"
                ],
                "summary": "Cập nhật dịch vụ",
                "operationId": "0db91423d9bab7ea76a82412abac1e20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy dịch vụ"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dich-vu/{id}/ngung-hoat-dong": {
            "put": {
                "tags": [
                    "Quản trị - Dịch vụ"
                ],
                "summary": "Ngừng hoạt động dịch vụ",
                "operationId": "fa72961650b6863fc694d1acde7dfc38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy dịch vụ"
                    },
                    "422": {
                        "description": "Không thể cập nhật trạng thái"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dich-vu/{id}/kich-hoat": {
            "put": {
                "tags": [
                    "Quản trị - Dịch vụ"
                ],
                "summary": "Kích hoạt lại dịch vụ",
                "operationId": "07b8d7e07d38ada3a8be56d4cb05e57a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy dịch vụ"
                    },
                    "422": {
                        "description": "Không thể cập nhật trạng thái"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}